requirements.txt
snowflake-connector-python
openai
flask
python-dotenv